moving the distributed folder to upper level
