+++ lmcache/v1/distributed/eviction_policy/__init__.py
@@ -4,10 +4,10 @@ Eviction policies for L1 cache management
 """
 
 # First Party
-from lmcache.v1.multiprocess.distributed.eviction_policy.factory import (
+from lmcache.v1.distributed.eviction_policy.factory import (
     CreateEvictionPolicy,
 )
-from lmcache.v1.multiprocess.distributed.eviction_policy.lru import (
+from lmcache.v1.distributed.eviction_policy.lru import (
     LRUEvictionPolicy,
 )
