+++ lmcache/v1/distributed/storage_controllers/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # First Party
-from lmcache.v1.multiprocess.distributed.storage_controllers.eviction_controller import (  # noqa: E501
+from lmcache.v1.distributed.storage_controllers.eviction_controller import (  # noqa: E501
     EvictionController,
 )
