+++ pyproject.toml
@@ -101,6 +101,7 @@ ignore = [
 [tool.ruff.lint.per-file-ignores]
 # Ignore SLF everywhere except lmcache/v1/multiprocess/
 "!lmcache/v1/multiprocess/**" = ["SLF"]
+"!lmcache/v1/distributed/**" = ["SLF"]
 
 [tool.ruff.lint.isort]
 # same as .isort.cfg
