+++ tests/v1/multiprocess/test_l1_manager.py
@@ -49,16 +49,16 @@ import pytest
 import torch
 
 # First Party
-from lmcache.v1.multiprocess.distributed.api import MemoryLayoutDesc, ObjectKey
-from lmcache.v1.multiprocess.distributed.config import (
+from lmcache.v1.distributed.api import MemoryLayoutDesc, ObjectKey
+from lmcache.v1.distributed.config import (
     L1ManagerConfig,
     L1MemoryManagerConfig,
 )
-from lmcache.v1.multiprocess.distributed.error import L1Error
+from lmcache.v1.distributed.error import L1Error
 
 try:
     # First Party
-    from lmcache.v1.multiprocess.distributed.l1_manager import L1Manager
+    from lmcache.v1.distributed.l1_manager import L1Manager
 except ImportError:
     # Skip tests if L1Manager cannot be imported
     pytest.skip(
